pyproject.toml
[project]
name = "excellent-insight"
version = "0.1.0"
description = "SaaS d'analyse automatique de fichiers Excel"
authors = [
    {name = "ExcellentInsight Team"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis[hiredis]>=5.2",
    "arq>=0.26",
    "polars>=1.20",
    "pyarrow>=18.0",
    "openpyxl>=3.1",
    "scipy>=1.14",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "email-validator>=2.2",
    "bcrypt>=4.2",
    "pydantic-settings>=2.7",
    "structlog>=24.4",
    "aiofiles>=24.1",
    "httpx>=0.28",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "pandas>=2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501", "E701", "E741", "B904", "F841", "F401", "E712", "B008", "F821", "F811"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
